Addition of src folder and real code
This commit adds the starting point for the real website hosted on github pages that will contain an about me introduction, links to related services, and the possibility of integrating Speech-To-Text to create interactive functionality. The src file introduces a concise directory for source code. Deprecated directory contains files that are not needed but still archived.